Fixed code with viewmodel
